AI-warehouse/requirements.txt
flask
flask-cors
gunicorn
pandas
numpy
xgboost
mysql-connector-python
duckdb
tabulate
scikit-learn
joblib
gunicorn==23.0.0